Kenzie/Sub-MAF-07: Refactored themes and Accessibility Settings Package.
