+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/Theme.Light"
-        tools:targetApi="31"
+    tools:targetApi="31"
         android:enableOnBackInvokedCallback="true"
         android:localeConfig="@xml/locales_config">
         <activity
@@ -42,7 +42,7 @@
             android:exported="true" />
 
         <activity
-            android:name="com.terracode.blueharvest.AccessibilitySettingsActivity"
+            android:name="com.terracode.blueharvest.accessibilitySettings.AccessibilitySettingsActivity"
             android:exported="true" />
 
     </application>
