+++ app/src/main/java/com/terracode/blueharvest/accessibilitySettings/AccessibilitySettingsActivity.kt
@@ -1,4 +1,4 @@
-package com.terracode.blueharvest
+package com.terracode.blueharvest.accessibilitySettings
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,43 +6,50 @@ import android.view.Menu
 import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
-import com.terracode.blueharvest.accessibilitySettings.AccessibilitySettingsActivity
+import com.terracode.blueharvest.ConfigurationSettingsActivity
+import com.terracode.blueharvest.HomeActivity
+import com.terracode.blueharvest.R
 
-class ConfigurationSettingsActivity : AppCompatActivity() {
-
+//Activity class for the accessibility setting page.
+class AccessibilitySettingsActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_configuration_settings)
+        setContentView(R.layout.activity_accessibility_settings)
 
         val toolbar: Toolbar = findViewById(R.id.settingsToolbar)
         setSupportActionBar(toolbar)
 
+        //Calls the accessibility settings from their respective classes
+        SetLanguage(this)
+        SetUnit(this)
     }
 
+    //Inflates the menu in the toolbar.
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
         menuInflater.inflate(R.menu.settings_menu, menu)
         return true
     }
 
+    //Logic for the different menu options (what activity to inflate).
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         return when (item.itemId) {
-            //This needs to be changed to include a card for notifications
             R.id.backButton -> {
                 val home = Intent(this, HomeActivity::class.java)
                 startActivity(home)
                 true
             }
+
             R.id.configurationSettings -> {
                 val operationSettings = Intent(this, ConfigurationSettingsActivity::class.java)
                 startActivity(operationSettings)
                 true
             }
+
             R.id.accessibilitySettings -> {
-                val accessibilitySettings = Intent(this, AccessibilitySettingsActivity::class.java)
-                startActivity(accessibilitySettings)
-                true
+                true // Do nothing, already in accessibility settings
             }
-            else -> false
+
+            else -> super.onOptionsItemSelected(item)
         }
     }
 }
