--- app/src/main/res/values-night/themes.xml
@@ -1,16 +1,0 @@
-<resources>
-    <!-- Base application theme. -->
-    <style name="Theme.MyApplication" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
-        <!-- Primary brand color. -->
-        <item name="colorPrimary">@color/darkOrange</item>
-        <item name="colorPrimaryVariant">@color/lightOrange</item>
-        <item name="colorOnPrimary">@color/black</item>
-        <!-- Secondary brand color. -->
-        <item name="colorSecondary">@color/blueGrey</item>
-        <item name="colorSecondaryVariant">@color/DarkGreen</item>
-        <item name="colorOnSecondary">@color/black</item>
-        <!-- Status bar color. -->
-        <item name="android:statusBarColor">@color/lavender</item>
-        <!-- Customize your theme here. -->
-    </style>
-</resources>
